requirements.txt
httpx==0.15.4
gpxpy==1.4.2
stravalib
appdirs>=1.4.0
svgwrite>=1.1.9
colour
s2sphere
arrow
geopy
polyline
sqlalchemy
timezonefinder
pyyaml
aiofiles
cloudscraper==1.2.58
python-tcxparser
rich
